reverted multi bean

+++ bean/bean.py
@@ -12,23 +12,24 @@ class Bean(commands.Cog):
     @commands.command()
     @commands.guild_only()
     @commands.has_permissions(manage_messages=True)
-    async def bean(self, ctx, targets: commands.Greedy[discord.Member], flags: str=None):
-        if flags != "-s":
-            for target in targets:
+    async def bean(self, ctx, target: discord.Member, flags: str=None):
+        if target.id != self.bot.user.id:
+            if flags != "-s":
                 await ctx.send('<:ItemBean:{}> Beaned **{}** (`{}`)'.format(self.beanEmoji.id, target, target.id))
-        else:
-            await ctx.message.delete()
+            else:
+                await ctx.message.delete()
 
-        try:
-            for target in targets:
-                message = await self.bot.wait_for('message', timeout=60*5, check=lambda m: m.author == target and m.channel.guild == ctx.guild)
-        except asyncio.TimeoutError:
-            pass
-        else:
             try:
-                await message.add_reaction(self.beanEmoji)
-            except Forbidden:
-                await message.channel.send('<:ItemBean:{}>'.format(self.beanEmoji.id))
+                message = await self.bot.wait_for('message', timeout=60*5, check=lambda m: m.author == target and m.channel.guild == ctx.guild)
+            except asyncio.TimeoutError:
+                pass
+            else:
+                try:
+                    await message.add_reaction(self.beanEmoji)
+                except Forbidden:
+                    await message.channel.send('<:ItemBean:{}>'.format(self.beanEmoji.id))
+        else:
+            await ctx.send('lol no.')
          
 
 def setup(bot):
